+++ spring-cloud-starter-tencent-polaris-config/src/test/java/com/tencent/cloud/polaris/config/listener/PolarisConfigRefreshOptimizationListenerTriggeredTest.java
@@ -21,7 +21,6 @@ package com.tencent.cloud.polaris.config.listener;
 import java.util.HashMap;
 import java.util.Map;
 
-import com.google.common.collect.Lists;
 import com.tencent.cloud.polaris.config.adapter.MockedConfigKVFile;
 import com.tencent.cloud.polaris.config.adapter.PolarisPropertySource;
 import com.tencent.cloud.polaris.config.adapter.PolarisPropertySourceManager;
@@ -31,6 +30,7 @@ import com.tencent.cloud.polaris.config.enums.RefreshType;
 import com.tencent.polaris.configuration.api.core.ChangeType;
 import com.tencent.polaris.configuration.api.core.ConfigKVFileChangeEvent;
 import com.tencent.polaris.configuration.api.core.ConfigPropertyChangeInfo;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
@@ -78,6 +78,11 @@ public class PolarisConfigRefreshOptimizationListenerTriggeredTest {
 	@Autowired
 	private ConfigurableApplicationContext context;
 
+	@BeforeEach
+	public void setUp() {
+		PolarisPropertySourceManager.clearPropertySources();
+	}
+
 	@Test
 	public void testSwitchConfigRefreshType() {
 		RefreshType actualRefreshType = context.getEnvironment()
@@ -101,12 +106,12 @@ public class PolarisConfigRefreshOptimizationListenerTriggeredTest {
 
 		PolarisPropertySource polarisPropertySource = new PolarisPropertySource(TEST_NAMESPACE, TEST_SERVICE_NAME, TEST_FILE_NAME,
 				file, content);
-		PolarisPropertySourceManager manager = context.getBean(PolarisPropertySourceManager.class);
-		when(manager.getAllPropertySources()).thenReturn(Lists.newArrayList(polarisPropertySource));
+		PolarisPropertySourceManager.addPropertySource(polarisPropertySource);
 
 		PolarisRefreshEntireContextRefresher refresher = context.getBean(PolarisRefreshEntireContextRefresher.class);
 		PolarisRefreshEntireContextRefresher spyRefresher = Mockito.spy(refresher);
 
+		refresher.setRegistered(false);
 		spyRefresher.onApplicationEvent(null);
 
 		ConfigPropertyChangeInfo changeInfo = new ConfigPropertyChangeInfo("k1", "v1", "v11", ChangeType.MODIFIED);
@@ -134,12 +139,6 @@ public class PolarisConfigRefreshOptimizationListenerTriggeredTest {
 
 	@SpringBootApplication
 	protected static class TestApplication {
-
-		@Primary
-		@Bean
-		public PolarisPropertySourceManager polarisPropertySourceManager() {
-			return mock(PolarisPropertySourceManager.class);
-		}
 
 		@Primary
 		@Bean